requirements.txt
# requirements.txt
csv 
requests 
mysql-connector-python
fpdf 
pandas 